Implement clone_extra_repos.py and unit tests (PHI Kanban #3050)
This script is mostly fully functional for cloning extra repos for a TriBITS
project.  All that is missing filtering repos based on gitolite for the
--gitolite-root option.  That will come later.

The basic functionality to read the extra repos from the file and filter them
base on type is complete.  Turns out most of the functionality I needed was
already implemented in the TriBITS CMake code!  That part was pretty easy.

Changings to existing code to support the clone_extra_repos.py script:

TribitsProcessExtraRepositoriesList.cmake:

* Added a new option ${PROJECT_NAME}_NO_CHECK_FOR_MISSING_EXTRA_REPOS.  This
is needed to get the list of repos when they don't yet exist (before we clone
them).  The default behavior is still to assert existance.

* Fixed a defect is setting ${PROJECT_NAME}_REPOSITORIES_CATEGORIES after
  filtering.  It was using a varible that was not even used in that macro.
  Macros are evil!

TribitsGetExtraReposForCheckinTest.cmake:

* Added NO_CHECK_FOR_MISSING_EXTRA_REPOS pass-threw.  This is used to get the
  list of repos before they exist.

* Change it so that when EXTRA_REPOS_PYTHON_OUT_FILE is not set, then the
  python dict will be printed to stdout/stderr instead of writing to a file.
  I don't want the clone_extra_repos.py script to create an intermediate file.