Skip C++11 checks and C++/Fortran managing checks if C++ disabled (Trilinos #6286)
With these changes, C++11 support processing is skipped if C++ is not even
enabled.  Also, we now skip the check for C++/Fortran interoperability if C++
is disabled.

These changes were made when I was testing the C-only build of Zoltan and
realized that if I turned of C++ then the C++11 checks were still being done
and so were the C++/Fortran managling checks (see Trilinos #6286).