Revert "Remove set of LD_LIBRARY_PATH now that using -Wl,-rpath (PHI Kanban #3217)"
This reverts commit 6d4f42cf2881df637b65189e0e1a631a26a81f34.

I was working in a build dir and it seems that adding the RPATH to the GCC lib
dir did not work for the executable xml2ctf.  Adding back the GCC lib dir to
LD_LIBRARY_PATH fixes it.  I wish I could find a robust way to get rpath to
work.