Fix disable of package tests/examples for TEST_REQUIRED_[PACKAGES,TPLS] disable
TriBITS was broken when a TEST_REQUIRED_TPLs TPL was disabled.  In that case,
it would hit ASSERT_DEFINED() for a var that is not defined yet in most cases.

TriBITS was also broken in a similar situration but for TEST_REQUIRED_PACKAGES
except in this case, it just would not actaully turn off the tests/examples
for the given package.

NOTE: It also turns out that if a variable is not defined, then ${VAR}
STREQUAL "" is FALSE *NOT* TRUE!  Therefore, you have to do "${${VAR}}"
STREQUAL "" to check this condition with possibily undefined vars.

I wrote unit tests to expose the problem then fixed the code.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_RELEASE_DEBUG_SHARED => passed: passed=327,notpassed=0 (2.51 min)