Add TRIBITS_INSTALL_HEADERS(), add HEADERS_INSTALL_SUBDIR arg (TriBITS #62, #55)
The main thing I did here was add the function TRIBITS_INSTALL_HEADERS().  It
turns out that this allowed me to aggregate some logic for installing header
files.  This can also be used to satisfy the request in TriBITS #55 to install
header files in multiple subdirs.  See the documentation for
TRIBITS_INSTALL_HEADERS() for details.

I also added the argument HEADERS_INSTALL_DIR to TRIBITS_ADD_LIBRARY() and
changed the TribitsExampleProject WithSubPackages subpackage 'C' to take
advantage of this.  There is a good unit test that makes sure this argument is
respected.  I had to adjust other stuff in TribitsExampleProject to compensate
for moving this header file C.hpp into a subdir.

The new TRIBITS_INSTALL_HEADERS() function is also now used to trim down the
code to install the ${REPOSITORY_NAME}_version.h file.  Someone reported a
problem with the installation of this file in Trilinos but I am not sure.