Fail hard when C++11 is enabled but support is not verified (TriBITS #64)
Before, when C++11 was enabled but support for C++11 could not be verified,
then C++11 was disbled.  Now that projects are shifting to requiring C++11,
the behavior of TriBITS is changed to now fail hard if C++11 support is turned
on but the configure-time tests fail.

Error messages are generated in one of two contexts.  The first context is
when the automatic check the C++11 compiler flags can't find flags that build
a test program.  The second context is when the C++11 compiler flags are
specified by the user but the configure-time compile checks for C++11 features
fails.

This fixes TriBITS issue #64.