Fix passing of TIMEOUT as arg not opt in TRIBITS_ADD_EXECUTABLE_AND_TEST() (TriBITS #65)
This fixes a defect in the handling of TIMEOUT in
TRIBITS_ADD_EXECUTABLE_AND_TEST().  It was handling it as an OPTION (i.e no
arguments).  There was a unit test but it was worthless because it was testing
TIMEOUT as an OPTION too!

I updated the unit test to pass in:

   TIMEOUT 11.5

instead of just

   TIMEOUT

and it showed the defect right away (see TriBITS #65).

It was then trival to fix the bug by changing TIMEOUT handling from an OPTION
to an ARGUMENT.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_RELEASE_DEBUG_SHARED => passed: passed=728,notpassed=0 (38.13 min)
Other local commits for this build/test group: ee77c81, 48da277