Fix path to win_interface/include and make conditional (TriBITS #26)
This commit fixes the path to the win_interface/interface that occured after
in the refactoring of the TriBITS repo in Github issue #26.  This issue was
discovered by John Cary (TechX).  There is no automated Windows testing
anymore with Trilinos so this was never seen in automated testing.

I also added an IF (EXISTS ...) statement to avoid the include if the
directory win_interface/include does not exist.  I image that most C++
projects will not need these hacks for Windows and therefore do not need to
snapshot the tribits/win_interface subdir.

I also added the varaible TRIBITS_SETUP_ENV_DEBUG to make it easier to debug
the happenings of TRIBITS_SETUP_ENV().

I tested this manually on Linux by manually configuring with -DMSVC=TRUE and
verified that the win_interface/include directory exists and shows up on the
build lines.