Remove usage of global var <Repo>_INSTALLED_REPO_VERSION_HEADER_FILE (Trilinos #6328)
Global varaibles are bad to use in general.  However, I don't believe there
was any defect in this code but the usage of this variable was confusing a
Trilinos developer (see Trilinos #6328).

To remove the usage of the global variable, the function
TRIBITS_REPOSITORY_CONFIGURE_ALL_VERSION_HEADER_FILES() now takes the required
argument ADD_INSTALL_TARGET.  This is called inside of TriBITS with 'TRUE'.
For the Trilinos usage that copies this file into the source tree under:

   Trilinos/SIERRA/bjam/config_headers/

this is passed 'FALSE'.

I manually tested this under Trilinos and it seems to behave correctly.