Remove warning for policy Policy CMP0054 for CMake 3.2.1
This should get rid of the warning:

CMake Warning (dev) at TriBITS/tribits/core/package_arch/TribitsAdjustPackageEnables.cmake:1327 (IF):
  Policy CMP0054 is not set: Only interpret if() arguments as variables or
  keywords when unquoted.  Run "cmake --help-policy CMP0054" for policy
  details.  Use the cmake_policy command to set the policy and suppress this
  warning.

  Quoted variables like "SCALE_ENABLE_TranscoreFields" will no longer be
  dereferenced when the policy is set to NEW.  Since the policy is not set
  the OLD behavior will be used.

Acutally this code was incorrect in that it should have dereferenced the
variable.  The unit tests fail if I comment out the clause that is fixed so
that is good enough testing for me.