Add basic support for pre-extra repos (TriBITS #73)
This is big commit.  There is a lot of refactorings to TriBITS to support
pre-extra repos.  This adds support for TriBITS Core.  Support for CI Support
and the CTest Driver will come later.  Also, more documentation is needed.

This commit changes the names of a bunch of internal variables but it should
not break backward compatibility seen by customer projects or users.

Some more specific changes/refactorings/additions include:

1) Replaced PACKSTAT with HASPKGS, REPOTYPE with VCTYPE: The new names are
more accurate and less confusing.  Now, 'PACKSTAT' is used for the field in
the 2D array that provides the HASPKS and PREPOST fields.  The term 'REPOTYPE'
is not clear.  The term 'VCTYPE' is clear and is what this field is used for.

To go with this, I also changeed the name of:

  ${PROJECT_NAME}_EXTRAREPOS_DIR_REPOTYPE_REPOURL_PACKSTAT_CATEGORY

to:

  ${PROJECT_NAME}_EXTRAREPOS_DIR_VCTYPE_REPOURL_PACKSTAT_CATEGORY

to be consistent.  But I maintain backward compatibility with the name
${PROJECT_NAME}_EXTRAREPOS_DIR_REPOTYPE_REPOURL_PACKSTAT_CATEGORY and issue a
deprecated warning.

2) Got rid of <Project>_NO_CHECK_FOR_MISSING_EXTRA_REPOS:

With just a little refactoring I was able to just use
${PROJECT_NAME}_CHECK_EXTRAREPOS_EXIST and simplify some logic.  One variable
has to be better than two.  It is very confusing otherwise.

3) Added TRIBITS_PARSE_EXTRAREPO_PACKSTAT() and unit tests: This is not a user
function but it makes it nice to unit test.

4) Changed TRIBITS_FILTER_OR_ASSERT_EXTRA_REPOS() from macro to function:
Macros have numerious namespace problems.  You should be explicit about what
varibles you are modifying in the caller's environment.

5) Print PRE/POST when adding extra repos: This is used be used for stronger
testing of pre-extra repos and just provides more detail.