CTest driver: Fix passing in list of extra repos separated with ','
This is needed to fix the 'make dashboard' target.  After the refactoring in
TriBITS #73, the 'make dashboard' target was broken.  It was only processing
the first POST extra repo.  This was due to the code not splitting the list of
extra repos coming from the 'make dashboard' target correctly.

We don't have any direct 'make dashboard' tests but I was able to add tests
for the TribitsCTestDriverCore.cmake file that exposed the defect and fix the
problem.  The fix was a simple split of the lists.

I also improved the printout of the ignore of NOPACKAGES repos a little.

Build/Test Cases Summary
Enabled Packages: TriBITS, CTeuchos, ForTeuchos
0) MPI_RELEASE_DEBUG_SHARED => passed: passed=202,notpassed=0 (3.99 min)