Document requirements of separate SE pkg source dirsA
There was a case in CASL VERA where the SCALE package 'Origen' subpackage
'DeprecatedIO' subdir was a subdir of the subpackage 'Manager' subdir.  In the
configure of VERA, the 'Manager' package was disabled and that added the
exclude 'Origen/Manager' which took out 'Origen/Manager/DeprecatedIO' which
broke the tarball.

This documentation update adds the explicit requirement that package dirs
can't nest.  That would avoid this problem.

 LocalWords:  unstage