Change from 'WARNING' to 'NOTE' to avoid CDash warning (TriBITS #80)
This should really be a 'NOTE' not a 'WARNING' anyway since it was requested
to ignore these directories.

I also improved the message about the disable of the required downstream
package.

This is technically breaking backward compatibility.  However, since this is
removing a bogous 'WARNING' that was being flagged in newer CDash versions, I
doubt that anywone is going to compain.