assert that HDF5 is found in its parallel version
As of recently, Debian supports installing both serial and MPI versions of HDF5
alongside. If both are installed, CMake prefers the serial version, leading to
funny errors further down in the compilation (unknown symbols for EpetraExt).
This commit makes sure that CMake bails out early if the serial version of HDF5
was found.