Correctly implement documented behavior of CDash email override (TriBITS #83)
The documented behavior of
${REPOSITORY_NAME}_REPOSITORY_OVERRIDE_PACKAGE_EMAIL_LIST was not being
implemented in the TriBITS CMake code.  While there was a way to use this
variable to override behavior, it was confusing and did not match documentation.

Now the usage of this variable matches documentation and is very
straightforward.

This is needed to allow VERA and MPACT to clearly override the CDash email
address used for Trilinos (Fixes TriBITS #83).