Add support for second CDash submit (PHI Kanban #3848)
This allows TRIBITS_CTEST_DRIVER() to optionally submit to a second CDash
site.  This is needed for the development of CDash so we can get submits and
all the data when implementing new features and to safely test an updated
version of CDash before switching over.

A few things I did as part of this:

* Refactored to use TRIBITS_CTEST_SUBMIT() wrapper and no longer overridding
  the default CTEST_SUBMIT() (I hate when people override basic langauge
  features).

* Add documentation for second CDash submits

* Updated list of vars that TribitsCTestDriverCore.cmake takes

Build/Test Cases Summary
Enabled Packages: TriBITS, CTeuchos
0) MPI_RELEASE_DEBUG_SHARED => passed: passed=200,notpassed=0 (6.44 min)