Add better gitdist-mod-status command
Here, all I really did was to change the repo ID's in the --dist-mod-only mode
to maintain the IDs when --dist-mod-only is not specified.  This makes it
easier to find the list of modified repos in the full list of repos table.

I therefore added the gitdist-mod-status alias which performs:

   gitdist --dist-mod-only dist-repo-status

Very useful I think.

Build/Test Cases Summary
Enabled Packages: TriBITS
Enabled all Forward Packages
0) MPI_RELEASE_DEBUG_SHARED => passed: passed=196,notpassed=0 (1.05 min)