Increase fetch timeout from 60 to 600 sec (PHI #3902)
The fetch was timing out for the SCALE repo when a bunch of binary data got
updated.  Hopefully increasing from 1 min to 10 min should do the trick.

I am not sure we even want to continue if there is a timeout because the
source dirs will be in an inconsistent state.

Also, we need to get an email if we detect any failure.  This needs to be
added to the TriBITS Dashboard Driver system.  We need to search for any
'CMake Error at' strings and make the outer test fail if we see any.

Build/Test Cases Summary
Enabled Packages: TriBITS
0) MPI_RELEASE_DEBUG_SHARED => passed: passed=197,notpassed=0 (0.95 min)