checkin-test.py: Remove suggestion to look at files in error message (TriBITS #93)
Before, checkin-test.py would try to guess what output files that users should
look at for more detail.  The problem is that was often wrong, throwing users
off.  Now it just prints the genetic message:

    To find out more about this failure, grep the 'checkin-test.out' log file
    for 'failed'.  In some cases, the failure will be obvious.  In other
    cases, a system command failed and the details about the failure will be
    in the output file for the command that failed.

That should be a bullet-proof way to find out what failed and why.

I added greps for this output in the unit tests to make sure it gets printed.

Fixes #93