Replace wrong/slow CMake-only impl with correct/fast cat/grep commands (PHI #3903)
The old CMake-only implementation was slow and consumed a *huge* amount of
memory with CMake 3.4.0 for a large CMakeCache.txt file.  It was also wrong in
the sense that it did not preserve ";" in list cache vars.

I did not bother trying to fix the CMake-only implementation.  I just added an
EXECUTE_PROCESS() command using cat and a few grep -v commands.  Now it runs
instantaneously.

I will next use this in TribitsCTestDriverCore.cmake and it should allow
upgrading to use CMake 3.4.0.