Fix passing of CTEST_PARALLEL_LEVEL for 'dashboard' target (PHI #3903)
The old code was passing '<N>' instead of <N> and CTEST_TEST() was ignoring it
because it did not convert to an integer.  This is CTest being "robust".
Sometimes I had "robust"!

Now it correctly runs the tests in parallel as you set using the CMake var
CTEST_PARALLEL_LEVEL.