Swap symlink/real file for gitdist.py/gitdist for Windows (TriBITS #98)
The clone_extra_repos.py script imports the gitdist.py module to get some of
its functionality.  On windows this fails because gitdist.py was just a
symlink to gitdist.  So, to support Windows, I just made the real file
gitdist.py and the symlink is now gitdist.  This should not impact Linux usage
at all.  And if you copy/install gitdist into another location, it will follow
the symlink and copy the file in its entirety.

With this change, on Windows, if you want to use gitdist, you will need to use
gitdist.py.

Fixes #98