Change to use splitlines() for Windows (TriBITS #98)
Changed from split("\n") to splitlines() to deal with newlines in MS Windows.

Jordan reports that this fixes a problem on MS Windows with the
clone_extra_repos.py script.  This also seems to be supported with Python
2.6.6 (which is what is running on the current machine).  This means that the
min version of Python required for TriBITS python support is now Python 2.6.6.

It actually took a good bit of work to adjust everything for this change as
split("\n") will always return an empty element at the end while splitlines()
will not.  It was tricky updating all of the code and tests due this simple
change.