Count as modified files with changes and stagged changes (TriBITS #15)
While this is not directly related to adding support for tracking branches
(TriBITS #15), I noticed that the function:

    gitdist.getNumModifiedAndUntracked()

was not counting as modified files with stagged and unstagged changes.  I read
the documentation for 'git status --porcelain' and saw all of the status
fields.  I added a unit test to check for all of the fields shown.

This now also marks and 'modified' files that are added 'A', files that are
renamed 'R', and unmerged files 'U' (i.e. files with merge conflict markers).