Use tracking branches and other improvements (TriBITS #15)
This is a significant refactoring of the checkin-test.py script to use
tracking branches for each repo independently and not assume 'origin' and not
assume that all of the repos are on the same branch or will be pulling and
pushing to the same remote branch.  This will make it easier to use the
checkin-test.py script to set up various integration scenarios.

Other advantages include:

1) You can now rename the primary remote from 'origin' or something more
descriptive like 'github' or 'casl-dev'.  This makes it eaier to keep things
stright when the workflow involves multiple remotes and so that you don't need
to keep running 'git remote -v' all the time to see what 'origin' is pointing
to.

2) You can set up integration scenarios where one or more of the repos points
to a different repo with a different tracking branch on fly without having to
hack the .git/config file to change what 'origin' points to.  Now you can just
checkout a different branch which points to a different tracking branch to a
different remote and then just run the checkin-test.py script.  Then, when you
are finished that integration sync, you can checkout your regular branch.

Other major improvements include:

1) On can now run the checkin-test.py script with the repos on a detached head
state using:

  --allow-no-pull/--local-do-all --enable-packages=<p0>,... --enable-all-pacakges=on/off

This allows you to use the checkin-test.py script to test versions while doing
a 'git bisect', for example.

2) Add print out of repo status table (from gitdist.py functions)

3) Removed a lot of extra git commands that don't need to be run.  For
example, if there are no commits w.r.t. the tracking branch, then there is no
need to do a 'git diff' w.r.t. the tracking branch.  Now there will not be any
git diffs at all if there are no commits for a repo.  Also, a
modifiedFiles.<repoDir>.out file will not even be generated if there are no
commits for a given repo.  That just makes the script run faster when you have
a lot of repos and it avoids clutter of the CHECKIN directory.

4) A push will be done for a repo if there are any commits but no diffs.  In
the past, if there was commits but no diff, then it would not push.  This is
rare but it has been a problem a few times over the years.

5) Trimmed down the list local commits included in final email and display
list even when no push.

Lower-level changes:

1) CheckinTest.py now depends on gitdist.py for several functions.  This is
how it gets the repos stats for each repo (tested against multiple versions of
git) and how it prints the repo status table (same as 'gitdist
dist-repo-status').  Some of the gitdist.py functions had to be refactored to
allow this but gitdist.py still does not depend on anything else.

2) The tests for checkin-test.py now grep for the final "REQUESTED ACTIONS:
PASSED" or "REQUESTED ACTIONS: FAILED" message to make sure the
checkin-test.py script completes successfully.  I added a new optional
argument grepForFinalPassFailStr for test cases where you want the script to
abort early.  I added this because I noticed that one of the tests reported as
passed because it was supposed to fail and all of the greps matched but the
script bombed in an unexpected way and did not complete gracefully like it was
supposed to.  Adding this grep by default makes sure that the script completes
gracefully by default.

3) Change from "update" to "pull" in printouts.  This is git so it is not
"update", it is "pull".

4) It turns out the existing test
test_extra_repo_file_3_continuous_do_all_push had a repo that had commits but
no diff.  I had to fix this test once I chagned the behavior as described
above.  To highlight this test case, I changed the name of the test to:

  test_extra_repo_file_3_continuous_commits_but_no_diff_do_all_push

Build/Test Cases Summary
Enabled Packages: TriBITS
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=199,notpassed=0 (0.98 min)
1) SERIAL_RELEASE => passed: passed=199,notpassed=0 (0.89 min)
Other local commits for this build/test group: 99fd955, cdeda00, 41f7110, 3d7567c, 8387998