Restore list of local commits when real push is done (TriBITS #15)
In my last commit, I had moved the code that gets the list of local commits
after the push.  That worked fine in the automated testing that uses fake git
commands.  But with a real git repo and real git commands:

  git log --oneline <branch> ^<tracking_branch>

returns no commits once the push is done.

When I did the real push, I noticed that the list of commits was empty.

This should fix this issue.

I also change the name g_cmnginterceptsEgLogCmnds to
g_cmnginterceptsGitLogCmnds.  Hopefully this is the last mention to 'eg'.

Build/Test Cases Summary
Enabled Packages: TriBITS
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=199,notpassed=0 (0.96 min)
1) SERIAL_RELEASE => passed: passed=199,notpassed=0 (0.81 min)