Allow current remote branch to have differnt names (TriBITS #15)
I accidentally did not fully update the push command.  I had it as:

   git push <remote> <trackingbranch>

But that assumes that the current branch is the same name as the tracking
branch.  That is not alway so.  Therefore, I have changed this to:

  git push <remote> <currentbranch>:<trackingbranch>

and I updated the automated tests to use different branch names to make sure
this works as it should.

This will make sure that no matter what the users settings are for 'git push'
or what version of git you are using, it will only push the current branch to
the tracking branch, period.

Build/Test Cases Summary
Enabled Packages: TriBITS
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=199,notpassed=0 (0.96 min)
1) SERIAL_RELEASE => passed: passed=199,notpassed=0 (0.80 min)