Fix defect in getCmndOutput() for throw
There was a defect in the throw block of the function getCmndOutput().  It
reference an undefined varible 'err' that should have been 'errCode'.  The
problem was that my unit test was just looking for an 'Exception' and it
bombed out without ever executing that branch.  I updated the unit test to
catch 'RuntimeError' and it exposed the problem.  I then fixed the defect.