Fix when repo has no tracking branch (TriBITS #15)
When the command:

   git rev-parse --abbrev-ref --symbolic-full-name @{u}

fails to return a tracking branch, it returns '128', not '0'.  The tests had
it returning '0'.  When I changed this command in these tests to correctly
return '128' when the repo does not have a tracking branch, it failed 7 of the
test cases.

The problem is that the gitdist.getCmndOutput() funtion never throwns an
exception but the GeneralScriptSupport.getCmndOutput() function does by
default.  Therefore, to pass in a version of
GeneralScriptSupport.getCmndOutput() that does not throw, i created the
function CheckinTest.GeneralScriptSupport.getCmndOutput() which calls
GeneralScriptSupport.getCmndOutput(cmnd, rtnCode=rtnCode, throwOnError=False).
With just that one change, everything passes.

Build/Test Cases Summary
Enabled Packages: TriBITS
Enabled all Forward Packages
0) MPI_RELEASE_DEBUG_SHARED => passed: passed=199,notpassed=0 (1.15 min)
Other local commits for this build/test group: a93d2d3