Make ambiguous HEAD obvious in output (TriBITS #100)
When I applied Chris's changes in his branch
cgbaker:filter_warnings_and_errors, it made my test 'test_all_ambiguous_head'
pass.  However, that version of the code just returned and empty field when
there in ambiguous HEAD.

Therefore, to make it more clear what the problem is, I made it return
'<AMBIGUOUS-HEAD>` in the table output to make it more clear that there is a
problem.

I validated this updated version of gitdist on a real git repo with an
ambiguous 'HEAD' created using:

  $ git tag HEAD

Fixes #100