Allow TRIBITS_CONFIGURE_CTEST_CUSTOM() to work for TDD system (PHI #3848)
Before, the TDD driver projects were not configuring the file
CTestCustom.cmake in the build dir.  This resulted in truncated test
output. This allows the TDD system to use the proper CTest settings, which it
needs for the long TribitsCTestDriverCore.cmake output logged to CDash as
tests.  For more details, see PHI #3848.

NOTE: There are no autoamted tests on the TDD side.  I had to just manaully
verify that it works with the TDD system.  But there is an automated TriBITS
test with the TribitsExampleProject example that makes sure this file gets
generated (or at least that the trace goes out).  I manually verified that
CTestCustom.cmake gets generated for VERA.