Create CDashQueryPassFail.py module (PHI #3936)
This is a completely generic Python module can be used with newer version of
CDash to query to see if builds for project are passing or not.

This will be used initailly to implement a 'deveop'/'master' branch workflow
for CASL VERA and then, perhaps, later for Trilinos.  A driver script must be
written for the particular project.  Once a driver script is completed for
Trilinos, one can see a concrete example.

Build/Test Cases Summary
Enabled Packages: TriBITS
Enabled all Forward Packages
0) MPI_RELEASE_DEBUG_SHARED => passed: passed=200,notpassed=0 (0.98 min)