Add CTEST_EXPLICITLY_ENABLE_IMPLICITLY_ENABLED_PACKAGES=FALSE (Trilinos #176)
As part of doing a mock run of a CTest driver as part of Trilinos #176, I had
to manaully set CTEST_EXPLICITLY_ENABLE_IMPLICITLY_ENABLED_PACKAGES=FALSE to
avoid processing all of the pakcages up to the package that I wanted to get
processed (MOOCHO).

I also reformatted the shell command to make it easier to see the options
being set.