If NetCDF is also enabled, may need hdf5_hl library
If using the netcdf4-enabled version of NetCDF library, it requires both libhdf5_hl and libhdf5.  

Check whether NetCDF is enabled and if so, add libhdf5_hl to the list of required libs.  This check should possibly also check if netcdf4 enabled in netcdf, but currently assumes that is the case.