Add -DNDEBUG to default RELEASE compiler flags, set Fortran flags
Raw CMake add -DNDEBUG to the default RELEASE compiler flags for C and C++.
This is what people seem to expect so that is what we should do.  Also, there
is an STK test that actually requires this (see trilinos/Trilinos#441).

This partially addresses trilinos/Trilinos#87.  To fully address that issue,
we must also these compiler flags for the Intel compiler as well.