Adjust to GCC 5.3.0 deprecated warning output
GCC 5.3.0 changed the formatting of its deprecated warnings.  This commit updates
the test that checks for this deprecated working to work with both versions of GCC.

NOTE: It is likely that this test will require more updates for more comiler
versions and we may even need to disable this test except for a few specific compilers.
On some of these other compilers, we might just look for the name of the deprecated functions
and "deprecated" to make sure that the compiler is showing a deprecated warning.
But I really like how strong these tests are right now with such specific output.