Clean up string conversion functions
The string conversion functions b(x), s(x) and u(x) were consolidated
and documented in the GeneralScriptSupport.py file (and in the
gitdist.py file, which is a stand-alone). Definitions of these
functions were removed from all of the other files that had them, and
they are now imported instead.