Fix echoRunSysCmnd() with extraEnv (#119)
The refactoring to make this Python code work with Python 2.6+ and 3.0+ broke
a print statement involving the extraEnv dict argument.  This broke automated
testing of Trilinos that uses the tdd_driver.py script which passes in an
extraEnv dict.  The script tdd_driver.py itself was never changed so it was
never looked at.  There is no automated testing for the tdd_driver.py script
(or most of the TriBITS Dashboard Driver functionality of TriBITS) so this was
not caught.

However, I was able to add a simple fast unit test for echoRunSyxCmnd()
passing in extraEnv and I was able to reproduce the failure.  I then updated
echoRunSysCmnd() so that it passes.

Hopefully this will allow the tdd_driver.py script to work correctly now.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=203,notpassed=0 (1.09 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)