Fix for errors when deleting a branch (PHI #3452)
When deleting ab branch, this script would print errors like:

  remote: fatal: not a tree object
  remote: ./hooks/pre-receive-block-large-files: line 71: [: -gt: unary operator expected

but it would remove the branch anyway.

With this fix, the script just aborts gracefully when a branch is deleted.
The trick is that git passes in 0000000000000000000000000000000000000000 for
the SHA1 of the updated branch.

I tested this manually and it works cleanly to push branches and delete
branches.