Make ALWAYS_FAIL_ON_NONZERO_RETURN print right pass/fail (#150)
It was doing the correct test case pass/fail but was not printing the right
pass/fail for this check.  Now it does.

I did not add a unit test for this since I think this is a more minor issue
(and I am not sure how to test it).

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=215,notpassed=0 (0.47 min)
1) SERIAL_RELEASE => passed: passed=215,notpassed=0 (0.43 min)