Make base Version.cmake file copy for Windows (#129)
You can't use a symlink for this file or it breaks the Windows build of a
TriBITS project (see #129).

I tested this manually a few different ways and it aborts the configure if it
detects that these two files are different.  Later, the Version.cmake file
will be configured so this will not be an issue.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=222,notpassed=0 (0.35 min)
1) SERIAL_RELEASE => passed: passed=222,notpassed=0 (0.31 min)