Set enable vars to 'OFF' for missing packages (#172)
This is needed to allow tarballs to work correctly when there are excluded
packages.  If you don't do this, then the configure of the still existing
downstream package may fail to build.

I put in pretty good unit tests for this.  I am pretty confident that this is
well demonstrated and protected.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=225,notpassed=0 (0.38 min)
1) SERIAL_RELEASE => passed: passed=225,notpassed=0 (0.39 min)