Fix bug in current behavior
If there are two libraries (e.g. CGNS and NetCDF) that both have a dependency
on the same library (e.g. HDF5), then the previous code would only add
the hdf5 dependencies to whichever of the two libraries were processed
first.

The current behavior fixes this by adding the dependency to both
libraries, but still only adds the target for the dependent library
once.