WIP: Use DISABLED test property for regular tests (#173)
Here, I updated the code for TRIBITS_ADD_TEST() to respond to
<test_name>_DISABLE=ON by setting the new DISABLED test property instead of
skipping the add of the test.

If have tested this manually with Trilinos/Teuchos and it sets the property
DISABLED and now the CTest version 3.8.20170206-gb9371 is marking this as "Not
Run" as a failing test locally!

This commit does not impact the TRIBITS_ADD_ADVANCED_TEST() function and the
unit tests for TriBITS have not been updated yet (they are currently failing).

This commit is just used to communicate with Kitware staff and allow use to
iterate on the proper functioning of this feature.