Set default for Fortran to OFF unconditionally (#190)
With the old setup, it was ignoring the setting of
${PROJECT_NAME}_ENABLE_Fortran_DEFAULT to OFF and was trying to find Fortran
anyway.  We found this while working on #190.

I added an automated test to make sure that TribitsHelloWorld (which sets the
default to OFF) is not trying to find Fortran.

This is a break in backward compatibility but about the only project affected
will be Trilinos.  When I snapshot TriBITS into Trilinos, I will add back the
logic to the Trilinos/CMakeLists.txt file to disable Fortran on Windows
(without Cygwin).

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=239,notpassed=0 (0.41 min)
1) SERIAL_RELEASE => passed: passed=239,notpassed=0 (0.36 min)