Replace PARSE_ARGUMENTS() with CMAKE_PARSE_ARGUMENTS() and deprecate (#181)
For the most part, this was a simple refactoring (just reorder the arguments
and put in an empty argument for the new one_value_keywords).  We did not use
the one_value_keywords argument because CMAKE_PARSE_ARGUMENTS() does not
actually validate that, it just ignores additional arguments (which is **NOT**
what we want).  We will need to add new functions to validate these arguments
instead :-(

In addition to changing the function call, CMAKE_PARSE_ARGUMENTS puts unparsed
arguments in a variable called #{PREFIX}_UNPARSED_ARGUMENTS PARSE_ARGUMENTS
however puts them in #{PREFIX}_DEFAULT_ARGS.  Some functions had to be changed
for this.

The test TribitsExampleProject_ALL_ST_NoFortran repeated TEST_5 and had to be
fixed.  With the PARSE_ARGUMENTS() implementation, it ignored the first TEST_5
block and used the second block.  With the CMAKE_PARSE_ARGUMENTS()
implementation, it gathered up the CMND arguments.  The new implementation
therefore caught the user error, but just did not report it very nicely.

The module ParseVariableArguments.cmake now generates a deprecated warning
when it is included and calling PARSE_ARGUMENTS() will generate a deprecated
warning every time you call it!

And this seems to improve configure time for CMake code using this by upwards
of 40% based on tests from TriBITS configures!  (The full configure time will
be less because of probing the env takes a lot of time.)

The test TribitsExampleProject_ALL_ST_NoFortran repeated TEST_5.  With the
PARSE_ARGUMENTS() implementation, it ignored the first TEST_5.  With the
CMAKE_PARSE_ARGUMENTS() implementation, it gathered up the CMND arguments.
The new implementation therefore caught the user error, but just did not
report it very nicely.