Fix path to CMake 2.8.11
Before, it was not picking up the path to CMake 2.8.11 and therefore was
testing with whatever (newer) version of CMake that was in the path.  The
Travis CI build caught this.  Just another example of how important post-push
CI testing is, even if you always have pre-push CI testing implemented.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=239,notpassed=0 (0.61 min)
1) SERIAL_RELEASE => passed: passed=239,notpassed=0 (0.63 min)