Fix defect in setting <Package>_ENABLE_<TPL>=ON (TriBITSPub/TriBITS#188)
This was a defect where I printed out that <Package>_ENABLE_<TPL>=ON would
*NOT* get set but then the code set it.  The automated test only checked for
the printout, not the actual value.  That was a bonehead defect that anyone
reviewing the code should have seen (including me).

It would be harder to write an automated test that actually checks the value
but I am in a hurry to fix this due to a defect this causes in Trilinos.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=239,notpassed=0 (0.93 min)
1) SERIAL_RELEASE => passed: passed=239,notpassed=0 (0.82 min)